refactor: Move UTCDateTime to server-utils.
