+++ robot-server/robot_server/persistence/tables/schema_03.py
@@ -2,7 +2,7 @@
 
 import sqlalchemy
 
-from robot_server.persistence._utc_datetime import UTCDateTime
+from server_utils.sql_utils import UTCDateTime
 
 metadata = sqlalchemy.MetaData()
