+++ robot-server/robot_server/persistence/tables/schema_10.py
@@ -4,7 +4,7 @@ import enum
 
 import sqlalchemy
 
-from robot_server.persistence._utc_datetime import UTCDateTime
+from server_utils.sql_utils import UTCDateTime
 
 metadata = sqlalchemy.MetaData()
